Simplify structure for the dialogs redux
