+++ src/App.js
@@ -25,10 +25,10 @@ import { createMuiTheme } from "@material-ui/core/styles";
 import { ThemeProvider } from "@material-ui/styles";
 import CssBaseline from "@material-ui/core/CssBaseline";
 
-import { userOperationsReducer } from "./Redux/userOperationsReducer";
-import { GlobalContext } from "./Redux/GlobalContext";
+// import { userOperationsReducer } from "./Redux/userOperationsReducer";
+// import { GlobalContext } from "./Redux/GlobalContext";
 
-import useCombinedReducers from "use-combined-reducers";
+// import useCombinedReducers from "use-combined-reducers";
 import EditProfileContainer from "./Containers/EditProfileContainer";
 
 import { SnackbarProvider } from "material-ui-snackbar-provider";
@@ -66,38 +66,38 @@ const theme = createMuiTheme({
 });
 
 const App = () => {
-  const [state, dispatch] = useCombinedReducers({
-    userOperations: useReducer(userOperationsReducer, {}),
-  });
+  // const [state, dispatch] = useCombinedReducers({
+  //   userOperations: useReducer(userOperationsReducer, {}),
+  // });
 
   return (
     <Provider store={store}>
-      <GlobalContext.Provider value={{ state, dispatch }}>
-        <ThemeProvider theme={theme}>
-          <SnackbarProvider SnackbarProps={{ autoHideDuration: 10000 }}>
-            <CssBaseline />
-            <BrowserRouter>
-              <Switch>
-                <Route exact={true} path={routes.HOME()} component={HomePage} />
-                <Route exact={true} path={routes.LIST_MEETUPS()} component={MeetupsContainer} />
-                <Route exact={true} path={routes.MEETUP_PAGE()} component={MeetupContainer} />
-                <PrivateRoute exact={true} path={routes.EVENT_SESSION_LIVE()} component={EventSessionContainer} />
-                <PrivateRoute exact={true} path={routes.EDIT_EVENT_SESSION()} component={EditSessionContainer} />
-                <Route exact={true} path={routes.EVENT_SESSION()} component={EventPageContainer} />
-                <Route exact={true} path={routes.LOGIN_PATH()} component={LoginContainer} />
-                <PrivateRoute path={routes.EDIT_PROFILE_RAW()} component={EditProfileContainer} />
-                <PrivateRoute path={routes.PROFILE()} component={ProfileContainer} />
-                <PrivateRoute path={routes.CREATE_EVENT_SESSION()} component={CreateSessionContainer} />
+      {/* <GlobalContext.Provider value={{ state, dispatch }}> */}
+      <ThemeProvider theme={theme}>
+        <SnackbarProvider SnackbarProps={{ autoHideDuration: 10000 }}>
+          <CssBaseline />
+          <BrowserRouter>
+            <Switch>
+              <Route exact={true} path={routes.HOME()} component={HomePage} />
+              <Route exact={true} path={routes.LIST_MEETUPS()} component={MeetupsContainer} />
+              <Route exact={true} path={routes.MEETUP_PAGE()} component={MeetupContainer} />
+              <PrivateRoute exact={true} path={routes.EVENT_SESSION_LIVE()} component={EventSessionContainer} />
+              <PrivateRoute exact={true} path={routes.EDIT_EVENT_SESSION()} component={EditSessionContainer} />
+              <Route exact={true} path={routes.EVENT_SESSION()} component={EventPageContainer} />
+              <Route exact={true} path={routes.LOGIN_PATH()} component={LoginContainer} />
+              <PrivateRoute path={routes.EDIT_PROFILE_RAW()} component={EditProfileContainer} />
+              <PrivateRoute path={routes.PROFILE()} component={ProfileContainer} />
+              <PrivateRoute path={routes.CREATE_EVENT_SESSION()} component={CreateSessionContainer} />
 
-                {/* <Route exact={true} path="/events" component={AllEventsContainer} /> */}
-                {/* <Route path="/event/:id" component={ViewEvent} />
+              {/* <Route exact={true} path="/events" component={AllEventsContainer} /> */}
+              {/* <Route path="/event/:id" component={ViewEvent} />
       <Route path="/newEvent" component={NewEvent} /> */}
-                <Route component={HomePage} />
-              </Switch>
-            </BrowserRouter>
-          </SnackbarProvider>
-        </ThemeProvider>
-      </GlobalContext.Provider>
+              <Route component={HomePage} />
+            </Switch>
+          </BrowserRouter>
+        </SnackbarProvider>
+      </ThemeProvider>
+      {/* </GlobalContext.Provider> */}
     </Provider>
   );
 };
