+++ src/Components/Event/ShareEventDialog.js
@@ -2,8 +2,7 @@ import React from "react";
 import Dialog from "@material-ui/core/Dialog";
 import { makeStyles } from "@material-ui/core/styles";
 import { useDispatch, useSelector } from "react-redux";
-import { closeShare } from "../../Redux/actions";
-import { isShareOpen } from "../../Redux/selectors";
+import { closeShare, isShareOpen } from "../../Redux/dialogs";
 import { getUrl } from "../../Modules/environments";
 import routes from "../../Config/routes";
 import { Typography, Tooltip, IconButton } from "@material-ui/core";
