--- src/Redux/GlobalContext.js
@@ -1,3 +1,0 @@
-import React from "react";
-
-export const GlobalContext = React.createContext();
