--- src/Redux/actionTypes.js
@@ -1,18 +1,0 @@
-export const OPEN_EDIT_PROFILE = "OPEN_EDIT_PROFILE";
-export const CLOSE_EDIT_PROFILE = "CLOSE_EDIT_PROFILE";
-
-export const OPEN_EVENT_DETAILS = "OPEN_EVENT_DETAILS";
-export const CLOSE_EVENT_DETAILS = "CLOSE_EVENT_DETAILS";
-
-export const OPEN_CHAT = "OPEN_CHAT";
-export const CLOSE_CHAT = "CLOSE_CHAT";
-export const CHAT_RESIZED = "CHAT_RESIZED";
-
-export const OPEN_SHARE = "OPEN_SHARE";
-export const CLOSE_SHARE = "CLOSE_SHARE";
-
-export const OPEN_FEEDBACK = "OPEN_FEEDBACK";
-export const CLOSE_FEEDBACK = "CLOSE_FEEDBACK";
-
-// export const SEND_CHAT_MESSAGE = "SEND_CHAT_MESSAGE";
-// export const CHAT_MESSAGE_SENT = "CHAT_MESSAGE_SENT";
