--- src/Redux/selectors.js
@@ -1,11 +1,0 @@
-export const isEditProfileOpen = (store) => store.dialogs.editProfileOpen === true;
-
-export const isEventDetailsOpen = (store) => store.dialogs.eventDetailsOpen === true;
-
-export const isChatOpen = (store) => store.dialogs.chatOpen === true;
-
-export const chatWidth = (store) => store.dialogs.chatWidth;
-
-export const isShareOpen = (store) => store.dialogs.shareOpen === true;
-
-export const isFeedbackOpen = (store) => store.dialogs.feedbackOpen === true;
